Name resolving seems overcomplicated #249
